alerts nos clientes

+++ lib/views/customers_view.dart
@@ -160,7 +160,68 @@ class CustomersView extends StatelessWidget {
                               icon: const Icon(Icons.delete),
                               color: const Color.fromARGB(255, 255, 0, 0),
                               onPressed: () {
-                                state.deleteCustomers(customer);
+                                showDialog(
+                                  context: context,
+                                  builder: (context) {
+                                    return AlertDialog(
+                                      title: const Padding(
+                                        padding: EdgeInsets.symmetric(
+                                          horizontal: 5,
+                                          vertical: 20,
+                                        ),
+                                        child: Text(
+                                          'Excluir Cliente',
+                                          style: TextStyle(
+                                            color: Colors.black,
+                                            fontWeight: FontWeight.bold,
+                                            fontSize: 25.0,
+                                          ),
+                                        ),
+                                      ),
+                                      content: const Text(
+                                        'Deseja excluir este cliente?',
+                                        style: TextStyle(
+                                          color: Colors.black,
+                                          fontSize: 20.0,
+                                        ),
+                                      ),
+                                      actions: [
+                                        TextButton(
+                                          onPressed: () {
+                                            Navigator.pop(context);
+                                          },
+                                          child: const Text(
+                                            'Cancelar',
+                                            style: TextStyle(
+                                              color: Colors.red,
+                                              fontWeight: FontWeight.bold,
+                                              fontSize: 20.0,
+                                            ),
+                                          ),
+                                        ),
+                                        TextButton(
+                                          onPressed: () {
+                                            state.deleteCustomers(customer);
+                                            Navigator.of(context).pop();
+                                          },
+                                          child: const Text(
+                                            'Confirmar',
+                                            style: TextStyle(
+                                              color: Colors.black,
+                                              fontSize: 20.0,
+                                            ),
+                                          ),
+                                        ),
+                                      ],
+                                      elevation: 25,
+                                      backgroundColor: Colors.white,
+                                      shape: RoundedRectangleBorder(
+                                        borderRadius:
+                                            BorderRadius.circular(50.0),
+                                      ),
+                                    );
+                                  },
+                                );
                               },
                             ),
                             IconButton(
